+++ src/model/formats/artworx.rs
@@ -44,7 +44,7 @@ pub fn read_adf(result: &mut Buffer, bytes: &[u8], file_size: usize) -> io::Resu
                 return Ok(true);
             }
             result.set_char(0, pos, Some(DosChar {
-                char_code: bytes[o],
+                char_code: bytes[o] as u16,
                 attribute: TextAttribute::from_u8(bytes[o + 1])
             }));
             pos.x += 1;
@@ -68,7 +68,7 @@ pub fn convert_to_adf(buf: &Buffer, options: &SaveOptions) -> io::Result<Vec<u8>
     for y in 0..buf.height {
         for x in 0..buf.width {
             let ch = buf.get_char(Position::from(x as i32, y as i32)).unwrap_or_default();
-            result.push(ch.char_code);
+            result.push(ch.char_code as u8);
             result.push(ch.attribute.as_u8());
         }
     }
