+++ src/model/tool/draw_ellipse_filled_imp.rs
@@ -12,7 +12,7 @@ pub struct DrawEllipseFilledTool {
     pub use_fore: bool,
     pub use_back: bool,
     pub attr: TextAttribute,
-    pub char_code: u8,
+    pub char_code: u16,
 }
 
 impl Plottable for DrawEllipseFilledTool {
@@ -25,7 +25,7 @@ impl Plottable for DrawEllipseFilledTool {
     fn get_use_back(&self) -> bool {
         self.use_back
     }
-    fn get_char_code(&self) -> u8 {
+    fn get_char_code(&self) -> u16 {
         self.char_code
     }
 }
